関数setservo()追加
